Suit Lock Config Rework
fixed the hyper beam config and made it indiscriminate (phazon is included too)
re-added suit bool for hyper beam usage
