+++ Content/SuitAddons/PhazonSuitAddon.cs
@@ -81,7 +81,7 @@ namespace MetroidMod.Content.SuitAddons
 			mp.EnergyDefenseEfficiency += 0.30f;
 			mp.EnergyExpenseEfficiency += 0.175f;
 			mp.phazonImmune = true;
-			mp.canUsePhazonBeam = true;
+			mp.accessPhazonBeam = true;
 			mp.UACost -= 0.10f;
 		}
 		public override void OnUpdateVanitySet(Player player)
